+++ src/main/kotlin/sparta/localconcert/domain/concerts/repository/RedisConcertRepository.kt
@@ -4,28 +4,14 @@ package sparta.localconcert.domain.concerts.repository
 import org.springframework.data.redis.connection.zset.Aggregate
 import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.stereotype.Repository
-import sparta.localconcert.global.config.ModuleConfig
+import sparta.localconcert.global.objectmapper.MapperConfig
 
 
 @Repository
 class RedisConcertRepository(
     private val redisTemplate: RedisTemplate<String, Any>,
-    private val objectMapper: ModuleConfig
+    private val objectMapper: MapperConfig
 ) {
-    fun <T> saveData(key: String, data: T) {
-        val mapper = objectMapper.objectMapper()
-        val value: String = mapper.writeValueAsString(data)
-        redisTemplate.opsForValue().set(key, value)
-    }
-
-    fun saveListData(key: String, data: List<Any>) {
-        val mapper = objectMapper.objectMapper()
-        for (element in data) {
-            val value: String = mapper.writeValueAsString(element)
-            redisTemplate.opsForList().leftPush(key, value)
-        }
-    }
-
     fun exists(key: String): Boolean {
         val result = redisTemplate.hasKey(key)
         return result
@@ -47,7 +33,6 @@ class RedisConcertRepository(
             redisTemplate.opsForZSet().incrementScore(key, value, 1.0)
         }
         searchKey()
-        println(searchKey())
         unionZSets()
     }
