+++ src/main/kotlin/sparta/localconcert/global/objectmapper/MapperConfig.kt
@@ -1,4 +1,4 @@
-package sparta.localconcert.global.config
+package sparta.localconcert.global.objectmapper
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Configuration
 
 
 @Configuration
-class ModuleConfig {
+class MapperConfig {
     @Bean
     fun objectMapper(): ObjectMapper {
         val objectMapper = ObjectMapper()
